moved more things to common
