+++ python/sst_module_program.py
@@ -1,63 +1,272 @@
 import os                
 import sst_module_common as sst_common
 import logging
-
+import re
 
 #####################################################################################
-## Given an array of RESULTS lines PER EVENT, generate the output file for this event
 #####################################################################################
-def create_output_file_results( output_dir_root: str, 
-                                event_num: int, 
-                                output_list: list, 
-                                display_relay_swimmer_names: bool,
-                                num_results_to_display: int ) -> int:
-    """ Generate the filename and open the next file """
+#####################################################################################
+#####################################################################################
+##########
+##########    P R O G R A M
+##########    process_program
+##########
+#####################################################################################
+#####################################################################################
+#####################################################################################
+#####################################################################################
+def process_program( meet_report_filename: str, 
+                     output_dir: str, 
+                     mm_license_name: str, 
+                     shorten_school_names_relays: bool, 
+                     shorten_school_names_individual: bool, 
+                     split_relays_to_multiple_files: bool, 
+                     add_new_line_to_relay_entries:bool, 
+                     display_relay_swimmer_names: bool,
+                     namesfirstlast: bool, 
+                    quote_output: bool ) -> int:
+    """ Given the input file formatted in a specific manner,
+        generate indiviual Event/Heat files for use in Wirecast displays """
     
+    #####################################################################################
+    ## The names are what appear in the report, and may be abbreviated, 
+    ##  and not the actual full school name
+    ## Multiple version of a school may be listed here for clean output
+    #####################################################################################
+
+    # program_relay_dict_full_name_len = 28
+    # program_ind_dict_full_name_len   = 25
+    # program_dive_dict_full_name_len  = 25    
+    # school_name_dict_short_name_len  = 4  
+
+    unofficial_results = ""  ## Not used for Programs
+
+    ## NOTE: Do not align up these headers with the TXT output.  
+    ##  Wirecast will center all lines and it will be in proper position then
+
+    program_header_len_dict = {
+        'individual_long': 25,
+        'diving_long': 25,
+        'relay_long': 28,
+    }
+
+    program_header_dict = {
+        'individual_long':   "\nLane  Name                    Year School      Seed Time",
+        'individual_short':  "\n  Lane  Name                   Yr Sch  Seed Time",
+        'diving_long':       "\nbLane  Name                 Year School      Seed Points",
+        'diving_short':      "\n  Lane  Name                     Yr Sch  Seed Points",
+        'relay_long':        "\nLane    Team                  Relay Seed Time" ,        
+        'relay_short':       "\nLane  Team Relay Seed Time",    
+    }
+
+    ## Define local variables
+    event_num = 0
+    heat_num = 0
     num_files_generated = 0
-    num_results_generated = 0
-    output_str = ""
+    num_header_lines = 3
+    found_header_line = 0
+    output_list = []
 
-    file_name_prefix = "results"
+    re_program_lane = re.compile('^[*]?\d{1,2} ')
+    re_program_lane_ind = re.compile('^(\d{1,2})\s+([A-z\' \.]+, [A-z ]+?) ([A-Z0-9]{1,2})\s+([A-Z \'.].*)\s+([X]?[0-9:.]+|NT|XNT|NP|XNP)*')
+    re_program_lane_relay = re.compile('^(\d{1,2})\s+([A-Z \'.].*)\s+([A-Z])\s+([X]?[0-9:.]+|NT|XNT)*')
 
-    output_dir = f"{output_dir_root}{file_name_prefix}/"
+    ## Remove the trailing hyphen (-) and -VA from school name
+    re_program_sch_cleanup1 = re.compile(r'(.*?)-VA(\s*)$')
+    re_program_sch_cleanup2 = re.compile(r'(.*?)-$(\s*)$')
 
-    ## Ignore the case where we get event0 heat0
-    if event_num == 0:
-        return 0
+    ## Search for headers to remove
+    # re_program_header_ind   = re.compile("^Lane(\s*)Name")
+    # re_program_header_relay = re.compile("^Lane(\s*)Team")
 
-    ## Create output dir if not exists
-    if not os.path.exists( output_dir ):
-        os.makedirs( output_dir )
+    ## Search for case where team time butts up against seed time.  
+    ## Need to add a space here so main regex works
+    re_program_space_team_seed = re.compile(r'([A-z])(X\d)')
 
-    ## Loop through list in reverse order
-    #for num in range( num_events-1, -1, -1):
-    for output_tuple in output_list:
-        row_type = output_tuple[0]
-        row_text = output_tuple[1]
+    ## For relays add a space between the persons name and next swimmer number
+    re_program_space_relay_name = re.compile(r'(\S)([2-4]\))')
+    re_program_check_relay_name_line = re.compile('1\)')
+    
+    ## Quote output for debugging
+    q = "'" if quote_output else ""
 
-        logging.info(f"RESULTS: e: {event_num} id: {row_type} t: {row_text}")
+    #####################################################################################
+    ## PROGRAM: Loop through each line of the input file
+    #####################################################################################
+    with open(meet_report_filename, "r") as meet_report_file:
+        for line in meet_report_file:
 
-        ## Save off the meet name, which somes at the end of the procesing as we are looping in reverse order
-        if row_type == 'H4':
-            output_str += row_text + '\n'
-            #output_str += '\n'
-        elif row_type == 'H6':
-            output_str += unofficial_results + '\n'
-            output_str += row_text + '\n'
-        elif row_type == 'PLACE':
-            output_str += row_text + '\n'
-            num_results_generated += 1
-        elif row_type == 'NAME' and display_relay_swimmer_names:
-            output_str += row_text + '\n'
+            #####################################################################################
+            ## PROGRAM: Remove the extra newline at end of line
+            #####################################################################################
+            line = line.strip()
 
-        if num_results_generated >= num_results_to_display:
-            break;
+            #####################################################################################
+            ## PROGRAM: Ignore all the blank lines             
+            #####################################################################################
+            if line == '\n' or line == '':
+                continue
 
-    output_file_name = output_dir + f"{file_name_prefix}_Event{event_num:0>2}.txt"
-    sst_common.write_output_file( output_file_name, output_str )
-    num_files_generated += 1
+            #####################################################################################
+            ## Meet Manager license name
+            ## We have one event/heat per page, so this starts the next event/heat
+            #####################################################################################
+            if re.search("^%s" % mm_license_name, line):
+                found_header_line = 1
+                
+                num_files = create_output_file_program( output_dir, event_num, heat_num, output_list, display_relay_swimmer_names, split_relays_to_multiple_files )
+                num_files_generated += num_files
 
+                ## Reset and start processing the next event/heat
+                output_list = []
+                output_list.append( ('H1', line ))
+                continue
+
+            #####################################################################################
+            ## if the previous line was the first header (found_header_line=1)
+            ## then save  the next two lines which are also part of the header and got to next line
+            #####################################################################################
+            if 0 < found_header_line < num_header_lines:
+                found_header_line += 1
+                if found_header_line == 2:
+                    output_list.append( ('H2', line ))
+                elif found_header_line == 3:
+                    output_list.append( ('H3', line ))
+                continue
+
+            #####################################################################################
+            ## Ignore these lines as we will add our own back in depending on the output format
+            #####################################################################################
+            ## For Relay Individual Events
+            # if re_program_header_ind.search(line):
+            #     continue
+            # ## For Relay Events
+            # if re_program_header_relay.search(line):
+            #     continue
+
+            #####################################################################################
+            ## PROGRAM: Start with Event line.  
+            ##  Get the Event Number from the report
+            ##  Clean it up
+            #####################################################################################
+            if line.lower().startswith(("event")):
+                event_num, event_str = sst_common.get_event_num_from_eventline( line )
+                ## H4 is the Event number/name line
+                output_list.append(('H4', f"{line} {unofficial_results}" ))
+
+
+            #####################################################################################
+            ## PROGRAM: Remove "Timed Finals" from Heat (and flight) line
+            #####################################################################################
+            if line.lower().startswith(("heat", "flight")):
+                line = line.replace("Timed Finals", "")
+                ## Remove all those extra spaces in the line
+                heat_num = sst_common.get_heat_num_from_heatline(line)
+
+                ## H6 is the Heat info, save it in case we want to output it later
+                output_list.append(('H5', f"{line}" ))
+
+                #####################################################################################
+                ## PROGRAM: Set name_list_header to be displayed above the list of swimmers
+                ##          This is only set once per Event/Heat so moving this is probablimetic
+                #####################################################################################
+                # Determin heading based on short or full school name
+                name_list_header = sst_common.get_header_line( event_num, shorten_school_names_relays, shorten_school_names_individual, program_header_dict ) 
+                if name_list_header != "":
+                    output_list.append(('H6', name_list_header))
+
+
+            #####################################################################################
+            ## PROGRAM: INDIVIDUAL Extract the individual Entry Line
+            ## i.e. 2   Robison, Ryan            JR  Bishop O'Connell-PV      X2:22.35                        
+            #####################################################################################
+            if (event_num in sst_common.event_num_individual or event_num in sst_common.event_num_diving) and re_program_lane.search(line):
+                ## Fix for case where School Name butts up to the X in seed time
+                line = re_program_space_team_seed.sub(r'\1 \2', line )
+
+                entry_line_list = re_program_lane_ind.findall(line)
+                #                                  LANE     LAST, FIRST        GR          SCHOOL             SEEDTIME 
+                if entry_line_list:
+                    entry_lane            = str(entry_line_list[0][0]).strip()
+                    entry_name_last_first = str(entry_line_list[0][1]).strip()
+                    entry_grade           = str(entry_line_list[0][2]).strip()
+                    entry_sch_long        = str(entry_line_list[0][3]).strip()
+                    entry_seedtime        = str(entry_line_list[0][4]).strip()
+                    
+                    ## If we want to use Shortened School Names, run the lookup
+                    if shorten_school_names_individual:
+                        ## The length of the school name in the MM report varies by event type
+                        school_name_len = program_header_len_dict['diving_long'] if event_num in sst_common.event_num_diving else program_header_len_dict['individual_long']
+
+                        entry_sch_short = sst_common.short_school_name_lookup( entry_sch_long, school_name_len )
+
+                    ## We can display name as given (Last, First) or change it to First Last with cli parameter
+                    entry_name = reverse_lastname_firstname( entry_name_last_first ) if namesfirstlast else entry_name_last_first
+
+                    ## Still issues with School names ending in - or -VA
+                    entry_sch_long = re_program_sch_cleanup1.sub(r'\1', entry_sch_long)
+                    entry_sch_long = re_program_sch_cleanup2.sub(r'\1', entry_sch_long)
+
+                    ## Format the output lines with either long (per meet program) or short school names
+                    output_str = f" {q}{entry_lane:>2}{q} {q}{entry_name:<25}{q} {q}{entry_grade:>2}{q} {q}{entry_sch_long:<25}{q} {q}{entry_seedtime:>8}{q}"
+
+                    if shorten_school_names_individual:
+                        output_str = f" {q}{entry_lane:>2}{q} {q}{entry_name:<25}{q} {q}{entry_grade:>2}{q} {q}{entry_sch_short:<4}{q} {q}{entry_seedtime:>8}{q}"
+                    
+                    output_list.append(('LANE', output_str))
+            
+            #####################################################################################
+            ## PROGRAM: RELAY Find the replay line with LANE, SCHOOL, RELAY TEAM SEEDTIME
+            ## 1 Seton Swim            A                    1:46.82      1:40.65        32
+            #####################################################################################
+            if event_num in sst_common.event_num_relay and re_program_lane.search(line):
+                entry_line_list = re_program_lane_relay.findall(line)
+                #  REGEX Positions                 LANE   SCHOOL    RELAY     SEEDTIME
+                if entry_line_list:
+                    entryline_place     = str(entry_line_list[0][0]).strip()
+                    entryline_sch_long  = str(entry_line_list[0][1]).strip()
+                    entryline_relay     = str(entry_line_list[0][2]).strip()
+                    entryline_seedtime  = str(entry_line_list[0][3]).strip()
+
+                    #####################################################################################
+                    ## PROGRAM: Replace long school name with short name for RELAY events
+                    #####################################################################################
+                    if shorten_school_names_relays:
+                        entryline_sch_short = entryline_sch_long
+                        entryline_sch_short = sst_common.short_school_name_lookup( entryline_sch_long, len(entryline_sch_long) )
+                        output_str = f"{q}{entryline_place:>2}{q} {q}{entryline_sch_short:<4}{q} {q}{entryline_relay:1}{q} {q}{entryline_seedtime:>8}{q}"
+                    else:
+                        ## Still issues with School names ending in - or -VA
+                        entryline_sch_long = re_program_sch_cleanup1.sub(r'\1', entryline_sch_long)
+                        entryline_sch_long = re_program_sch_cleanup2.sub(r'\1', entryline_sch_long)
+
+                        output_str = f"{q}{entryline_place:>2}{q} {q}{entryline_sch_long:<28}{q} {q}{entryline_relay:1}{q} {q}{entryline_seedtime:>8}{q}"
+
+                    output_list.append(( "LANE", output_str ))
+
+            #####################################################################################
+            ## PROGRAM: RELAY Add the swimmers name to the list. It may or may not be use for output
+            #####################################################################################
+            ## If this is a relay, add a space between the last swimmer name and the next swimmer number
+            ## This line  1) LastName1, All2) LastName2, Ashley3) LastName3, All4) LastName4, Eri
+            ## becomes    1) LastName1, All 2) LastName2, Ashley 3) LastName3, All 4) LastName4, Eri
+            if event_num in sst_common.event_num_relay and re_program_check_relay_name_line.search(line):
+                output_str = re_program_space_relay_name.sub( r'\1 \2',line )
+                output_list.append(( "NAME", output_str ))
+
+    #####################################################################################
+    ## Reached end of file
+    ## Write out last event
+    #####################################################################################
+
+    num_files = create_output_file_program( output_dir, event_num, heat_num, output_list, display_relay_swimmer_names, split_relays_to_multiple_files )
+    num_files_generated += num_files
+
+    #####################################################################################
+    ## PROGRAM: All done. Return counts of files created
+    #####################################################################################
     return num_files_generated
+
